Getting closer to fully functional App
Now need the Sensors, NewSensor, and Connector functions to populate
the indexes and funtions to search the indexes and return specific
sensors.

Might need to defer the ability to use strings for volt and amp
methods and require users to use the function names initially.
Just check to make sure that the function is not nil.