Add an "every N times" read option
Add a factory for creating a "do real read only every N times" read
function.  Might be useful for v1.0 hadware when reading voltages
on channels 9-15 which a user noticed were slow with v1.0 software
due to the averaging of 16 values.

NOTE: It's kinda cool, but these can be composited.  The readfn
to everyn_factory can be the output of filter_factory!

Also fix a mistake where we failed to export to pi the mcp3008 and
bbwain _read functions