Add a pidev_setup() function to the pidev API
Add a _setup function so users of the library can set the global
variables before calling pidev_open to set the location of the
libexecdir and config file and specify the ARGV0 string.

The pointers passed must correspond to allocated storage that
will not be freed.  pidev doesn't make a copy of the strings.

Passing NULL for any of those string arguments leaves them at
the defaults (libexecdir = ".", configfile = "pidev.conf")

Also set the starting debug and verbose values.