Add new temperature sensor type.  44004 standard thermistor
Add resistance to temp and temp to resistance functions.
Export them in several places to provide user access to them.

This sensor uses a special cable with a built-in resistor to connect
the thermistor between 5V and GND and the "Amp" input pin on a J (volt,amps)
connector.  This provides a ratiometric read of the resistance relative
to the resistor in the cable with no influence from the Vcc rail.